Makefile: Add update-embedded-root rule
This uses the "plumbing" command to ensure the newest root has been
downloaded and verified. Then it copies the newest TUF root and the
trusted_root.json into the sources. The benefit here is that one does
not need to manually find the cache directories when an update should
be done.

This hard codes XDG_DATA_HOME and XDG_CACHE_HOME for simplicity.

We could later add a workflow that runs this on cron and files an
issue if the sources changed as a result.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>